requirements.txt
fastapi
uvicorn
pydantic  
sqlalchemy
passlib==1.7.4
bcrypt==4.0.1
python-jose